Use a typedef for monetary values

+++ src/qt/transactiontablemodel.cpp
@@ -536,7 +536,7 @@ QVariant TransactionTableModel::data(const QModelIndex &index, int role) const
         case ToAddress:
             return formatTxToAddress(rec, true);
         case Amount:
-            return rec->credit + rec->debit;
+            return static_cast<qint64>(rec->credit + rec->debit);
         }
         break;
     case Qt::ToolTipRole:
@@ -569,7 +569,7 @@ QVariant TransactionTableModel::data(const QModelIndex &index, int role) const
     case LabelRole:
         return walletModel->getAddressTableModel()->labelForAddress(QString::fromStdString(rec->address));
     case AmountRole:
-        return rec->credit + rec->debit;
+        return static_cast<qint64>(rec->credit + rec->debit);
     case TxIDRole:
         return rec->getTxID();
     case TxHashRole:
